Support for start stop downsampler

+++ src/tools/CliQuery.java
@@ -205,14 +205,14 @@ final class CliQuery {
         i++;
         
         long counterMax = Long.MAX_VALUE;
-        long resetValue = RateOptions.DEFAULT_RESET_VALUE;
+        double resetValue = RateOptions.DEFAULT_RESET_VALUE;
         if (args[i].startsWith("counter")) {
           String[] parts = Tags.splitString(args[i], ',');
           if (parts.length >= 2 && parts[1].length() > 0) {
             counterMax = Long.parseLong(parts[1]);
           }
           if (parts.length >= 3 && parts[2].length() > 0) {
-            resetValue = Long.parseLong(parts[2]);
+            resetValue = Double.parseDouble(parts[2]);
           }
           rate_options = new RateOptions(true, counterMax, resetValue);
           i++;
